refactor: remove files and change style
