+++ src/App.css
@@ -1,5 +1,4 @@
 .main-title {  
-  /* color: rgb(21,21,40); */
   color: rgb(249,148,2);
   text-align: center;
   padding: 1rem;
