+++ src/components/card/card.module.css
@@ -39,7 +39,7 @@
 }
 
 .name {
-font-weight: 800;
+font-weight: 700;
 text-transform: uppercase;
 margin-bottom: 1rem;
 color: rgb(249,148,2);
