+++ src/components/header/header.module.css
@@ -2,7 +2,6 @@
   display: flex;
   justify-content: space-around;
   flex: 0 0 auto;
-  /* background: rgb(140, 140, 141); */
   background: rgb(86, 123, 149);
 }
