+++ src/components/nav/nav.module.css
@@ -14,7 +14,6 @@
   color: white;
   text-decoration: none;
   padding: 1rem;
-  /* background: rgb(113, 113, 118); */
   background-color: rgb(63, 100, 127);
   cursor: default;
 }
