--- src/components/svgComponent/svgComponent.tsx
@@ -1,9 +1,0 @@
-import { FC } from 'react';
-
-interface SVGProp {
-  src: string;
-  id: number;
-}
-export const SVGComponent: FC<SVGProp> = ({ src, id }) => {
-  return <img src={src} id={id.toString()} />;
-};
