--- src/dataCards.ts
@@ -1,185 +1,0 @@
-import { CardProps } from './components/card/Card';
-
-import img1 from '/src/assets/img/1.webp';
-import img2 from '/src/assets/img/2.webp';
-import img3 from '/src/assets/img/3.webp';
-import img4 from '/src/assets/img/4.webp';
-import img5 from '/src/assets/img/5.webp';
-import img6 from '/src/assets/img/6.webp';
-import img7 from '/src/assets/img/7.webp';
-import img8 from '/src/assets/img/8.webp';
-import img9 from '/src/assets/img/9.webp';
-import img10 from '/src/assets/img/10.webp';
-import img11 from '/src/assets/img/11.webp';
-import img12 from '/src/assets/img/12.webp';
-import img13 from '/src/assets/img/13.webp';
-import img14 from '/src/assets/img/14.webp';
-import img15 from '/src/assets/img/15.webp';
-import img16 from '/src/assets/img/16.webp';
-import img17 from '/src/assets/img/17.webp';
-import img18 from '/src/assets/img/18.webp';
-import img19 from '/src/assets/img/19.webp';
-import img20 from '/src/assets/img/20.webp';
-
-export const dataCards: CardProps[] = [
-  {
-    id: 1,
-    src: img1,
-    title: 'maqi LAVA COOKIES',
-    author: 'ka Lok Hong',
-    likes: 199,
-    show: 2000,
-  },
-  {
-    id: 2,
-    src: img2,
-    title: 'Collection 2020',
-    author: 'SATISH GANGAIAH',
-    likes: 380,
-    show: 4700,
-  },
-  {
-    id: 3,
-    src: img3,
-    title: "'Trobada' Wall, Cantonigròs (Catalonia)",
-    author: 'Reskate Studio',
-    likes: 102,
-    show: 748,
-  },
-  {
-    id: 4,
-    src: img4,
-    title: 'Anton Muhin PH & Sasha',
-    author: 'anga make up',
-    likes: 38,
-    show: 414,
-  },
-  {
-    id: 5,
-    src: img5,
-    title: 'KLM - Up For It',
-    author: 'Woodwork Amsterdam',
-    likes: 538,
-    show: 5300,
-  },
-  {
-    id: 6,
-    src: img6,
-    title: 'Voice recorder',
-    author: 'Layo Studio',
-    likes: 739,
-    show: 7700,
-  },
-  {
-    id: 7,
-    src: img7,
-    title: 'Collection Character vol1',
-    author: 'Yossef Salama',
-    likes: 223,
-    show: 1300,
-  },
-  {
-    id: 8,
-    src: img8,
-    title: 'A Night At The Opera for Tatler UK',
-    author: 'Velweiss Retouch',
-    likes: 74,
-    show: 654,
-  },
-  {
-    id: 9,
-    src: img9,
-    title: 'The Face Art Academy',
-    author: 'Egle Ruibyte',
-    likes: 478,
-    show: 610,
-  },
-  {
-    id: 10,
-    src: img10,
-    title: 'Winds of Change',
-    author: 'Allan Peters',
-    likes: 387,
-    show: 3200,
-  },
-  {
-    id: 11,
-    src: img11,
-    title: 'Jaime Hayde Artbook',
-    author: 'Jaime Hayde',
-    likes: 337,
-    show: 3200,
-  },
-  {
-    id: 12,
-    src: img12,
-    title: 'Porsche at Night',
-    author: 'GF Williams',
-    likes: 160,
-    show: 2000,
-  },
-  {
-    id: 13,
-    src: img13,
-    title: 'Niwaki Catalogue',
-    author: 'Sophie Chatellier',
-    likes: 168,
-    show: 2000,
-  },
-  {
-    id: 14,
-    src: img14,
-    title: 'TYPE 0.7',
-    author: 'Huston Wilson',
-    likes: 484,
-    show: 3000,
-  },
-  {
-    id: 15,
-    src: img15,
-    title: 'Festival Packaging for WAGAS',
-    author: 'Inkee Wang',
-    likes: 730,
-    show: 5000,
-  },
-  {
-    id: 16,
-    src: img16,
-    title: 'MAMA',
-    author: 'FAENA Studio',
-    likes: 425,
-    show: 2900,
-  },
-  {
-    id: 17,
-    src: img17,
-    title: 'XKR_002 | Protect',
-    author: 'XK stydio',
-    likes: 542,
-    show: 2800,
-  },
-  {
-    id: 18,
-    src: img18,
-    title: 'Mooah #003',
-    author: 'Hanee O',
-    likes: 135,
-    show: 1100,
-  },
-  {
-    id: 19,
-    src: img19,
-    title: 'peep coffee',
-    author: 'Hui Tong',
-    likes: 5400,
-    show: 47700,
-  },
-  {
-    id: 20,
-    src: img20,
-    title: 'The cozy purrs',
-    author: 'Đốm Đốm',
-    likes: 5400,
-    show: 47700,
-  },
-];
