+++ src/index.css
@@ -17,7 +17,6 @@ body {
   padding: 0;
   font-size: 16px;
   color: rgb(21, 21, 40);
-  /* background: rgb(232, 232, 237); */
   background: rgb(242,239,220);
   display: flex;
   flex-direction: column;
